# Conflicts:
#	.gitignore
#	main.py
#	utils.py

+++ main.py
@@ -23,6 +23,8 @@
 # SOFTWARE.
 
 import argparse
+import multiprocessing
+import time
 import warnings
 from typing import Any
 from pathlib import Path
@@ -38,6 +40,8 @@ from torch import nn, Tensor
 from torchvision import transforms
 from torch.utils.data import DataLoader
 
+from models import *
+import utils
 from dataset import SliceDataset
 from ShallowNet import shallowCNN
 from ENet import ENet
@@ -56,6 +60,9 @@ from utils import (Dcm,
                    visualize_sample,
                    save_images)
 
+from losses import create_loss_fn
+
+import wandb #TODO: remove all wandb instances on final submission
 from losses import (CrossEntropy)
 import matplotlib.pyplot as plt
 
@@ -66,19 +73,16 @@ datasets_params["TOY2"] = {'K': 2, 'net': shallowCNN, 'B': 2}
 datasets_params["SEGTHOR"] = {'K': 5, 'net': ENet, 'B': 8}
 
 
-def setup(args) -> tuple[nn.Module, Any, Any, DataLoader, DataLoader, int]:
+def setup(args) -> tuple[nn.Module, torch.optim.Optimizer, torch.optim.lr_scheduler.LRScheduler, Any, DataLoader, DataLoader, int]:
     # Networks and scheduler
     gpu: bool = args.gpu and torch.cuda.is_available()
     device = torch.device("cuda") if gpu else torch.device("cpu")
     print(f">> Picked {device} to run experiments")
 
     K: int = datasets_params[args.dataset]['K']
-    net = datasets_params[args.dataset]['net'](1, K)
+    net = eval(args.model)(1, K, **vars(args))
     net.init_weights()
     net.to(device)
-
-    lr = 0.0005
-    optimizer = torch.optim.Adam(net.parameters(), lr=lr, betas=(0.9, 0.999))
 
     # Dataset part
     B: int = datasets_params[args.dataset]['B']
@@ -107,6 +111,10 @@ def setup(args) -> tuple[nn.Module, Any, Any, DataLoader, DataLoader, int]:
         itemgetter(0)
     ])
 
+    # Used to seed dataloader workers, passed in `generator` param
+    g = torch.Generator()
+    g.manual_seed(args.seed)
+
     train_set = SliceDataset('train',
                              root_dir,
                              img_transform=img_transform,
@@ -116,7 +124,9 @@ def setup(args) -> tuple[nn.Module, Any, Any, DataLoader, DataLoader, int]:
     train_loader = DataLoader(train_set,
                               batch_size=B,
                               num_workers=args.num_workers,
-                              shuffle=True)
+                              shuffle=True,
+                              worker_init_fn=utils.seed_worker,
+                              generator=g)
 
     val_set = SliceDataset('val',
                            root_dir,
@@ -127,11 +137,19 @@ def setup(args) -> tuple[nn.Module, Any, Any, DataLoader, DataLoader, int]:
     val_loader = DataLoader(val_set,
                             batch_size=B,
                             num_workers=args.num_workers,
-                            shuffle=False)
+                            shuffle=False,
+                            worker_init_fn=utils.seed_worker,
+                            generator=g)
 
     args.dest.mkdir(parents=True, exist_ok=True)
 
-    return (net, optimizer, device, train_loader, val_loader, K)
+    # lr = 0.0005 # Initial LR for ENet
+    lr = args.lr
+    optimizer = torch.optim.AdamW(net.parameters(), lr=lr, betas=(0.9, 0.999), weight_decay=args.lr_weight_decay)
+    scheduler = None
+    if args.enable_lr_scheduler:
+        scheduler = torch.optim.lr_scheduler.OneCycleLR(optimizer, lr, epochs=args.epochs, steps_per_epoch=len(train_loader))
+    return net, optimizer, scheduler, device, train_loader, val_loader, K
 
 #TODO Decide necessity of this?
 def skip_empty_masks(gt: Tensor, pred_seg: Tensor) -> bool:
@@ -143,21 +161,20 @@ def skip_empty_masks(gt: Tensor, pred_seg: Tensor) -> bool:
 
 
 def runTraining(args):
-    print(f">>> Setting up to train on {args.dataset} with {args.mode}")
-    net, optimizer, device, train_loader, val_loader, K = setup(args)
 
-    if args.mode == "full":
-        loss_fn = CrossEntropy(idk=list(range(K)))  # Supervise both background and foreground
-    elif args.mode in ["partial"] and args.dataset in ['SEGTHOR', 'SEGTHOR_STUDENTS']:
-        loss_fn = CrossEntropy(idk=[0, 1, 3, 4])  # Do not supervise the heart (class 2)
-    else:
-        raise ValueError(args.mode, args.dataset)
+    utils.seed_everything(args)
+    start = time.time()
+    print(f">>> Setting up to train on {args.dataset} with {args.model}")
+    net, optimizer, scheduler, device, train_loader, val_loader, K = setup(args)
+
+    loss_fn = create_loss_fn(args, K)
 
     # Notice one has the length of the _loader_, and the other one of the _dataset_
     log_loss_tra: Tensor = torch.zeros((args.epochs, len(train_loader)))
     log_dice_tra: Tensor = torch.zeros((args.epochs, len(train_loader.dataset), K))
     log_loss_val: Tensor = torch.zeros((args.epochs, len(val_loader)))
     log_dice_val: Tensor = torch.zeros((args.epochs, len(val_loader.dataset), K))
+
     log_jacc_tra: Tensor = torch.zeros((args.epochs, len(train_loader.dataset), K))
     log_jacc_val: Tensor = torch.zeros((args.epochs, len(val_loader.dataset), K))
     log_ahd_tra: Tensor = torch.zeros((args.epochs, len(train_loader.dataset), K))
@@ -170,6 +187,7 @@ def runTraining(args):
     log_recall_val: Tensor = torch.zeros((args.epochs, len(val_loader.dataset), K))
 
     best_dice: float = 0
+    best_metrics = {}
 
     for e in range(args.epochs):
         for m in ['train', 'val']:
@@ -245,7 +263,7 @@ def runTraining(args):
 
                         # Computing AHD per class for each batch
                         ahd_values_per_class = average_hausdorff_distance_per_class(gt[batch_idx], pred_seg[batch_idx], K)
-                        for k in range(K): 
+                        for k in range(K):
                             if ahd_values_per_class[k] != float('inf'):  # Skipping 'inf' values
                                 log_ahd[e, j + batch_idx, k] = ahd_values_per_class[k]
 
@@ -262,8 +280,11 @@ def runTraining(args):
                     if opt:  # Only for training
                         loss.backward()
                         opt.step()
+                        # Apply LR scheduler
+                        if scheduler:
+                            scheduler.step()
 
-                    if m == 'val':
+                    if m == 'val' and not args.dry_run:
                         with warnings.catch_warnings():
                             warnings.filterwarnings('ignore', category=UserWarning)
                             predicted_class: Tensor = probs2class(pred_probs)
@@ -289,18 +310,23 @@ def runTraining(args):
                                          for k in range(1, K)}
                         postfix_dict |= {f"AHD-{k}": f"{log_ahd[e, :j, k].mean():05.3f}"
                                          for k in range(1, K)}
-                        postfix_dict |= {f"Precision-{k}": f"{log_precision[e, :j, k].mean():05.3f}" 
+                        postfix_dict |= {f"Precision-{k}": f"{log_precision[e, :j, k].mean():05.3f}"
                                         for k in range(1, K)}
-                        postfix_dict |= {f"Recall-{k}": f"{log_recall[e, :j, k].mean():05.3f}" 
+                        postfix_dict |= {f"Recall-{k}": f"{log_recall[e, :j, k].mean():05.3f}"
                                         for k in range(1, K)}
                     tq_iter.set_postfix(postfix_dict)
 
-                #Adding these printed Metrics when Job is running as sanity check
-                print(f"Epoch {e} - {m} AHD: {log_ahd[e, :j, 1:].mean().item():05.3f}")
-                print(f"Epoch {e} - {m} Jaccard: {log_jacc[e, :j, 1:].mean().item():.5f}")
-                print(f"Epoch {e} - {m} ASSD: {log_assd[e, :j].mean().item():.5f}")
-                print(f"Epoch {e} - {m} Precision: {log_precision[e, :j, 1:].mean().item():.5f}")
-                print(f"Epoch {e} - {m} Recall: {log_recall[e, :j, 1:].mean().item():.5f}")
+
+        metrics = utils.save_loss_and_metrics(K, e, args.dest,
+                                              loss=[log_loss_tra, log_loss_val],
+                                              dice=[log_dice_tra, log_dice_val])
+        wandb.log(metrics)
+        #Adding these printed Metrics when Job is running as sanity check
+        print(f"Epoch {e} - {m} AHD: {log_ahd[e, :j, 1:].mean().item():05.3f}")
+        print(f"Epoch {e} - {m} Jaccard: {log_jacc[e, :j, 1:].mean().item():.5f}")
+        print(f"Epoch {e} - {m} ASSD: {log_assd[e, :j].mean().item():.5f}")
+        print(f"Epoch {e} - {m} Precision: {log_precision[e, :j, 1:].mean().item():.5f}")
+        print(f"Epoch {e} - {m} Recall: {log_recall[e, :j, 1:].mean().item():.5f}")
 
         # I save it at each epochs, in case the code crashes or I decide to stop it early
         np.save(args.dest / "loss_tra.npy", log_loss_tra)
@@ -327,23 +353,31 @@ def runTraining(args):
             with open(args.dest / "best_epoch.txt", 'w') as f:
                     f.write(str(e))
 
-            best_folder = args.dest / "best_epoch"
-            if best_folder.exists():
-                    rmtree(best_folder)
-            copytree(args.dest / f"iter{e:03d}", Path(best_folder))
+            if not args.dry_run:
+                best_folder = args.dest / "best_epoch"
+                if best_folder.exists():
+                        rmtree(best_folder)
+                copytree(args.dest / f"iter{e:03d}", Path(best_folder))
 
             torch.save(net, args.dest / "bestmodel.pkl")
             torch.save(net.state_dict(), args.dest / "bestweights.pt")
+            best_metrics = metrics
+
+    for key, value in best_metrics.items():
+        wandb.run.summary[key] = value
+    end = time.time()
+    print(f"[FINISHED] Duration: {(end - start):0.2f} s")
 
 
 def main():
     parser = argparse.ArgumentParser()
 
     parser.add_argument('--epochs', default=200, type=int)
-    parser.add_argument('--dataset', default='TOY2', choices=datasets_params.keys())
+    parser.add_argument('--dataset', default='SEGTHOR', choices=datasets_params.keys())
     parser.add_argument('--mode', default='full', choices=['partial', 'full'])
     parser.add_argument('--dest', type=Path, required=True,
                         help="Destination directory to save the results (predictions and weights).")
+    parser.add_argument('--seed', default=42, type=int, help='Random seed to use for reproducibility of the experiments')
 
     parser.add_argument('--num_workers', type=int, default=5)
     parser.add_argument('--gpu', action='store_true')
@@ -351,13 +385,52 @@ def main():
                         help="Keep only a fraction (10 samples) of the datasets, "
                              "to test the logic around epochs and logging easily.")
 
+    parser.add_argument('--dropoutRate', type=float, default=0.2, help="Dropout rate for the ENet model")
+    parser.add_argument('--lr', type=float, default=0.0005, help="Learning rate")
+    parser.add_argument('--lr_weight_decay', type=float, default=0.1, help="Weight decay factor for the AdamW optimizer")
+    parser.add_argument('--enable_lr_scheduler', action='store_true')
+
+    parser.add_argument('--alpha', type=float, default=0.5, help="Alpha parameter for loss functions")
+    parser.add_argument('--beta', type=float, default=0.5, help="Beta parameter for loss functions")
+    parser.add_argument('--focal_alpha', type=float, default=0.25, help="Alpha parameter for Focal Loss")
+    parser.add_argument('--focal_gamma', type=float, default=2.0, help="Gamma parameter for Focal Loss")
+
+    # Optimize snellius batch job
     parser.add_argument('--scratch', action='store_true', help="Use the scratch folder of snellius")
+    parser.add_argument('--dry_run', action='store_true', help="Disable saving the image validation results on every epoch")
+    parser.add_argument('--disable_wandb', action='store_true', help="Disable the WandB logging")
+
+    # Arguments for more flexibility of the run
     parser.add_argument('--remove_unannotated', action='store_true', help="Remove the unannotated images")
+    parser.add_argument('--loss', default='CrossEntropy', choices=['CrossEntropy', 'Dice', 'FocalLoss', 'CombinedLoss', 'FocalDiceLoss', 'TverskyLoss'])
+    parser.add_argument('--model', type=str, default='ENet', choices=['ENet', 'shallowCNN', 'UNet', 'UNetPlusPlus', 'DeepLabV3Plus'])
+    parser.add_argument('--run_prefix', type=str, default='', help='Name to prepend to the run name')
+    parser.add_argument('--run_group', type=str, default=None, help='Your name so that the run can be grouped by it')
+
+    # Arguments for running with different backbones
+    parser.add_argument('--encoder_name', type=str, default='resnet18', choices=['resnet18', 'resnet34', 'resnet50', 'resnet101', 'resnet152'])
+    parser.add_argument('--unfreeze_enc_last_n_layers', type=int, default=1, help="Train the last n layers of the encoder")
 
     args = parser.parse_args()
+    run_name = utils.get_run_name(args, parser)
+    args.dest = args.dest / run_name
 
+    # Added since for python 3.8+, OS X multiprocessing starts processes with spawn instead of fork
+    # see https://github.com/pytest-dev/pytest-flask/issues/104
+    multiprocessing.set_start_method("fork")
+
+    utils.wandb_login(args.disable_wandb)
+    wandb.init(
+        entity="ai_4_mi",
+        project="SegTHOR",
+        name=run_name,
+        config=vars(args),
+        mode="disabled" if args.disable_wandb else "online",
+        group=args.run_group
+    )
+
+    print(f">> {run_name} <<")
     pprint(args)
-
     runTraining(args)
